update GenerateGroup params timestamp to long

+++ src/main/java/org/fisco/bcos/web3j/protocol/core/Ethereum.java
@@ -98,7 +98,7 @@ public interface Ethereum {
     Request<?, SendTransaction> sendRawTransaction(String signedTransactionData);
 
     // generateGroup
-    Request<?, GenerateGroup> generateGroup(int groupId, int timestamp, List<String> nodeList);
+    Request<?, GenerateGroup> generateGroup(int groupId, long timestamp, List<String> nodeList);
 
     Request<?, StartGroup> startGroup(int groupId);
